+++ components/linker/props.ts
@@ -1,4 +1,4 @@
-import { booleanProp, buildProps, eventProp } from '@vexip-ui/config'
+import { booleanProp, buildProps, eventProp, iconProp } from '@vexip-ui/config'
 
 import type { ExtractPropTypes, PropType } from 'vue'
 import type { ConfigurableProps } from '@vexip-ui/config'
@@ -7,7 +7,7 @@ import type { LinkerType } from './symbol'
 export const linkerProps = buildProps({
   to: String,
   type: String as PropType<LinkerType>,
-  icon: Object,
+  icon: iconProp,
   underline: booleanProp,
   disabled: booleanProp,
   target: String,