+++ docs/demos/number-input/sync/demo.en-US.vue
@@ -7,6 +7,15 @@
     :max="10"
     style="max-width: 300px"
   ></NumberInput>
+  <p>Only sync step change:</p>
+  <NumberInput
+    v-model:value="value"
+    sync-step
+    clearable
+    :min="-10"
+    :max="10"
+    style="max-width: 300px"
+  ></NumberInput>
   <p>Number Input Value: {{ value }}</p>
 </template>