+++ common/utils/CHANGELOG.md
@@ -1,3 +1,36 @@
+# [2.10.0](https://github.com/vexip-ui/vexip-ui/compare/utils@2.9.0...utils@2.10.0) (2024-01-04)
+
+
+### ✨ Features
+
+* **utils:** add decide method ([87f2f06](https://github.com/vexip-ui/vexip-ui/commit/87f2f06fc5eae45b12816e591add212e204feaf9))
+
+
+### 🐞 Bug Fixes
+
+* **utils:** dom util methods consider whether is client ([81c0ce3](https://github.com/vexip-ui/vexip-ui/commit/81c0ce32ed7f932677d81d3804e40d0cb99e4d90))
+
+
+
+# [2.9.0](https://github.com/vexip-ui/vexip-ui/compare/utils@2.8.0...utils@2.9.0) (2023-12-27)
+
+
+### ✨ Features
+
+* **utils:** callIfFunc support provide extra function parameters ([6653ff7](https://github.com/vexip-ui/vexip-ui/commit/6653ff7b12cfe24032383c762ea474eb30933df0))
+
+
+### 🐞 Bug Fixes
+
+* **utils:** isValidNumber should consider scientific notation ([72edf02](https://github.com/vexip-ui/vexip-ui/commit/72edf025c5d555f30bc67d7facd720a34bc1f4ff))
+
+
+### 🔨 Code Refactoring
+
+* **utils:** rename transformListToMap to listToMap ([cc40220](https://github.com/vexip-ui/vexip-ui/commit/cc4022075fce035a3eaa45be8fafddb771ecbc1f))
+
+
+
 # [2.8.0](https://github.com/vexip-ui/vexip-ui/compare/utils@2.7.3...utils@2.8.0) (2023-12-19)